+++ src/scanner/enumerate.rs
@@ -144,7 +144,7 @@ pub fn enumerate_filesystem_inputs(
         Some(&matcher_stats),
         enable_profiling,
         Some(shared_profiler),
-        args.compat_ignore_comments,
+        &args.extra_ignore_comments,
         args.no_inline_ignore,
     )?;
     let blob_processor_init_time = Mutex::new(t1.elapsed());