+++ src/scanner/repos.rs
@@ -622,7 +622,7 @@ pub async fn fetch_s3_objects(
         Some(matcher_stats),
         enable_profiling,
         Some(shared_profiler.clone()),
-        args.compat_ignore_comments,
+        &args.extra_ignore_comments,
         args.no_inline_ignore,
     )?;
     let mut processor = BlobProcessor { matcher };